Resolve missing blockarray docs
