+++ docs/source/classes.rst
@@ -6,6 +6,7 @@ Important Classes
 .. toctree::
    :maxdepth: 1
 
+   blockarray
    operator
    functional
    optimizer
